Creating download manager to determine when to download file
