+++ app/src/main/java/com/duckduckgo/app/browser/downloader/FileDownloader.kt
@@ -19,13 +19,14 @@ package com.duckduckgo.app.browser.downloader
 import android.os.Environment
 import android.webkit.URLUtil
 import androidx.annotation.WorkerThread
-import com.duckduckgo.app.browser.downloader.NetworkFileDownloader.UserDownloadAction
+import com.duckduckgo.app.browser.downloader.NetworkFileDownloadManager.DownloadFileData
+import com.duckduckgo.app.browser.downloader.NetworkFileDownloadManager.UserDownloadAction
 import java.io.File
 import javax.inject.Inject
 
 class FileDownloader @Inject constructor(
     private val dataUriDownloader: DataUriDownloader,
-    private val networkDownloader: NetworkFileDownloader
+    private val networkFileDownloadManager: NetworkFileDownloadManager
 ) {
 
     @WorkerThread
@@ -36,9 +37,9 @@ class FileDownloader @Inject constructor(
         }
 
         when {
-            URLUtil.isNetworkUrl(pending.url) -> networkDownloader.download(pending, callback)
+            URLUtil.isNetworkUrl(pending.url) -> networkFileDownloadManager.download(pending, callback)
             URLUtil.isDataUrl(pending.url) -> dataUriDownloader.download(pending, callback)
-            else -> callback?.downloadFailed("Not supported")
+            else -> callback.downloadFailed("Not supported")
         }
     }
 
@@ -52,7 +53,7 @@ class FileDownloader @Inject constructor(
 
     interface FileDownloadListener {
         fun confirmDownload(
-            downloadFileData: NetworkFileDownloader.DownloadFileData,
+            downloadFileData: DownloadFileData,
             userDownloadAction: UserDownloadAction
         )
         fun downloadStarted()
