+++ app/src/main/java/com/duckduckgo/app/browser/downloader/NetworkFileDownloader.kt
@@ -21,42 +21,14 @@ import android.content.Context
 import android.webkit.CookieManager
 import android.webkit.URLUtil
 import androidx.core.net.toUri
-import com.duckduckgo.app.browser.downloader.FileDownloader.FileDownloadListener
 import timber.log.Timber
-import java.io.File
 import javax.inject.Inject
 
 class NetworkFileDownloader @Inject constructor(private val context: Context) {
 
-    fun download(
-        pendingDownload: FileDownloader.PendingFileDownload,
-        callback: FileDownloadListener
-    ) {
+    fun download(pendingDownload: FileDownloader.PendingFileDownload) {
         val guessedFileName = guessFileName(pendingDownload)
-        val fileToDownload = File(pendingDownload.directory, guessedFileName)
-        val alreadyDownloaded = fileToDownload.exists()
-        callback.confirmDownload(
-            DownloadFileData(fileToDownload, alreadyDownloaded),
-            object : UserDownloadAction {
-                override fun acceptAndReplace() {
-                    File(pendingDownload.directory, guessedFileName).delete()
-                    initiateDownload(pendingDownload, guessedFileName)
-                }
 
-                override fun accept() {
-                    initiateDownload(pendingDownload, guessedFileName)
-                }
-
-                override fun cancel() {
-                    Timber.i("Cancelled download for url ${pendingDownload.url}")
-                }
-            })
-    }
-
-    private fun initiateDownload(
-        pendingDownload: FileDownloader.PendingFileDownload,
-        guessedFileName: String?
-    ) {
         val request = DownloadManager.Request(pendingDownload.url.toUri()).apply {
             allowScanningByMediaScanner()
             addRequestHeader("Cookie", CookieManager.getInstance().getCookie(pendingDownload.url))
@@ -67,18 +39,9 @@ class NetworkFileDownloader @Inject constructor(private val context: Context) {
         manager?.enqueue(request)
     }
 
-    private fun guessFileName(pending: FileDownloader.PendingFileDownload): String {
-        val guessedFileName =
-            URLUtil.guessFileName(pending.url, pending.contentDisposition, pending.mimeType)
+    private fun guessFileName(pending: FileDownloader.PendingFileDownload): String? {
+        val guessedFileName = URLUtil.guessFileName(pending.url, pending.contentDisposition, pending.mimeType)
         Timber.i("Guessed filename of $guessedFileName for url ${pending.url}")
         return guessedFileName
     }
-
-    interface UserDownloadAction {
-        fun accept()
-        fun acceptAndReplace()
-        fun cancel()
-    }
-
-    class DownloadFileData(val file: File, val alreadyDownloaded: Boolean)
 }
